Rename init.js to lib/register.js
