+++ lib/register.js
@@ -1,6 +1,6 @@
-const lint = require('./lib/lint')
-const { GRAMMAR_SCOPES: grammarScopes } = require('./lib/lint')
-const cleanLinters = require('./lib/getLinter').cleanLinters
+const lint = require('./lint')
+const { GRAMMAR_SCOPES: grammarScopes } = require('./lint')
+const cleanLinters = require('./getLinter').cleanLinters
 
 exports.deactivate = () => {
   cleanLinters()
